[💄Design, 🚚Rename]: 공통된 css 수정, 폴더 이름 변경 1. main.css에 있던 공통으로 쓰이는 css 여기 폴더로 수정 2. 기능구현중 너무 어려워서 일단 밑에다가 참고 기능 주석으로 추가 3. 폴더이름 구분되게 수정
